Refactores some functions into a helpers module
